requirements.txt
anthropic>=0.18.0
pillow>=10.0.0
pillow-heif>=0.13.0
